Closes #2807: move tab opening to MainActivity
