+++ app/src/main/java/org/mozilla/tv/firefox/webrender/WebRenderFragment.kt
@@ -14,9 +14,6 @@ import android.view.View
 import android.view.ViewGroup
 import android.view.WindowManager
 import android.widget.FrameLayout
-import androidx.lifecycle.Lifecycle
-import io.reactivex.Observable
-import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
 import io.reactivex.rxkotlin.addTo
@@ -39,7 +36,6 @@ import org.mozilla.tv.firefox.MediaSessionHolder
 import org.mozilla.tv.firefox.R
 import org.mozilla.tv.firefox.ScreenControllerStateMachine.ActiveScreen
 import org.mozilla.tv.firefox.architecture.FirefoxViewModelProviders
-import org.mozilla.tv.firefox.ext.application
 import org.mozilla.tv.firefox.ext.couldScrollInDirection
 import org.mozilla.tv.firefox.ext.focusedDOMElement
 import org.mozilla.tv.firefox.ext.isYoutubeTV
@@ -49,16 +45,12 @@ import org.mozilla.tv.firefox.ext.requireWebRenderComponents
 import org.mozilla.tv.firefox.ext.resetView
 import org.mozilla.tv.firefox.ext.scrollByClamped
 import org.mozilla.tv.firefox.ext.serviceLocator
-import org.mozilla.tv.firefox.ext.startMainActivity
 import org.mozilla.tv.firefox.ext.updateFullscreenScrollPosition
 import org.mozilla.tv.firefox.ext.webRenderComponents
-import org.mozilla.tv.firefox.fxa.FxaReceivedTab
 import org.mozilla.tv.firefox.hint.HintBinder
 import org.mozilla.tv.firefox.hint.InactiveHintViewModel
 import org.mozilla.tv.firefox.session.SessionRepo
 import org.mozilla.tv.firefox.utils.URLs
-import org.mozilla.tv.firefox.utils.ViewUtils
-import java.util.concurrent.TimeUnit
 
 private const val ARGUMENT_SESSION_UUID = "sessionUUID"
 
@@ -80,7 +72,6 @@ class WebRenderFragment : EngineViewLifecycleFragment(), Session.Observer {
     private val mediaSessionHolder get() = activity as MediaSessionHolder? // null when not attached.
 
     private val startStopCompositeDisposable = CompositeDisposable()
-    private val createDestroyCompositeDisposable = CompositeDisposable()
 
     // If YouTubeBackHandler is instantiated without an EngineView, YouTube won't
     // work properly, so we !!
@@ -94,9 +85,6 @@ class WebRenderFragment : EngineViewLifecycleFragment(), Session.Observer {
         initSession()
 
         webRenderViewModel = FirefoxViewModelProviders.of(this).get(WebRenderViewModel::class.java)
-
-        observeReceivedTabs()
-            .addTo(createDestroyCompositeDisposable)
     }
 
     @SuppressLint("RestrictedApi")
@@ -233,8 +221,6 @@ class WebRenderFragment : EngineViewLifecycleFragment(), Session.Observer {
 
         HintBinder.bindHintsToView(hintViewModel, hintBarContainer, animate = true)
                 .forEach { startStopCompositeDisposable.add(it) }
-
-        maybeOpenQueuedTab()?.addTo(startStopCompositeDisposable)
     }
 
     override fun onStop() {
@@ -242,11 +228,6 @@ class WebRenderFragment : EngineViewLifecycleFragment(), Session.Observer {
 
         serviceLocator!!.sessionRepo.exitFullScreenIfPossible()
         startStopCompositeDisposable.clear()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        createDestroyCompositeDisposable.clear()
     }
 
     override fun onDestroyView() {
@@ -272,41 +253,6 @@ class WebRenderFragment : EngineViewLifecycleFragment(), Session.Observer {
                         viewToFocus.requestFocus()
                     }
                 }
-    }
-
-    private fun maybeOpenQueuedTab(): Disposable? {
-        val queuedTab = serviceLocator!!.fxaRepo.queuedFxaTabs.poll() ?: return null
-
-        // This is a really bad solution, and could cause future race conditions. Executing
-        // immediately throws "java.lang.IllegalStateException: FragmentManager is already
-        // executing transactions", and I couldn't find any better way to deconflict
-        return Observable.just("unused")
-            .delay(1, TimeUnit.SECONDS)
-            .observeOn(AndroidSchedulers.mainThread())
-            .subscribe { openReceivedFxaTab(queuedTab) }
-    }
-
-    private fun observeReceivedTabs(): Disposable {
-        return serviceLocator!!.fxaRepo.receivedTabs.subscribe {
-            val appIsInForeground = lifecycle.currentState == Lifecycle.State.STARTED ||
-                lifecycle.currentState == Lifecycle.State.RESUMED
-            val appMightBeInBackground = lifecycle.currentState == Lifecycle.State.CREATED
-
-            when {
-                appIsInForeground -> openReceivedFxaTab(it)
-                appMightBeInBackground -> {
-                    serviceLocator!!.fxaRepo.lastTabCouldNotBeDisplayed()
-                    context?.application?.startMainActivity()
-                }
-                else -> { }
-            }
-        }
-    }
-
-    private fun openReceivedFxaTab(receivedTab: FxaReceivedTab) {
-        // TODO: Gracefully handle receiving multiple tabs around the same time. #2777
-        serviceLocator!!.screenController.showBrowserScreenForUrl(fragmentManager!!, receivedTab.url)
-        ViewUtils.showCenteredBottomToast(context, receivedTab.tabReceivedNotification.resolve(resources))
     }
 
     fun loadUrl(url: String) {
